backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
werkzeug==3.1.3
pandas==3.0.1
numpy==2.4.4
